\name{asent_setup}
\alias{asent_setup}
\title{Set up Python environment}
\usage{
asent_setup(python, first = FALSE)
}
\description{
Set up Python environment and enable automatic download of pre-trained model.
}
\arguments{
\item{python}{A character vector specifying the location of Python.}

\item{model}{A logical value indicating whether to download pre-trained model. Default is \code{FALSE}.}
}
\value{
A list of functions from {asent}
}
